Fix bugs from Bugs6.md
Emulator:
- alloc_space() range checks the refs-to-bytes scaling before malloc, so
  a --size-heap of a multiple of 2^61 refs no longer wraps to a
  zero-byte block that the VM then treats as 16 exabytes and never
  collects (H-1).
- The GC's new-space-exhausted guards and the two illegal-instruction
  default: labels are no longer compiled out by --enable-ndebug: they
  cost a pointer comparison and two jump table entries, and they are
  what turns a corrupt world or a GC sizing bug into a diagnostic
  instead of a no-op or an out-of-bounds write.  The per-instruction
  stack buffer check stays behind FAST; it measures about 30% of the
  interpreter's running time on tak (M-2).
- --size-seg-max is bounded so that 255 + max_segment_size cannot
  overflow int and skip the stack sizing guarantee altogether (M-4).
- Cold worlds carry an oakcold32/oakcold64 word size marker, which
  read_world checks, so a cold world of the wrong word size is refused
  rather than crashing somewhere unrelated.  oak-cold-linker and
  tool.oak both write it (M-5).
- oak-cold-linker consults errno after strtoll and range checks
  tagize_int, and tool.oak's tagize-int makes the matching check, so an
  integer constant too big for the target's fixnum is reported rather
  than quietly linked as a different number (M-3).
- GET-ARGLINE-CHAR takes its indices as ssize_t, so an index congruent
  to a valid one mod 2^32 no longer aliases onto it (L-2); the stack
  buffer byte count is computed in size_t (L-4); the threaded build's
  default newspace size comes from DEFAULT_NEWSPACE like the other one
  (L-6); and SIGINT is installed with sigaction rather than depending on
  the platform's signal() semantics (L-7).

World:
- The string writers mask the character code they deposit to its own
  byte.  A code above 255 was shifting its high bits into the byte of
  the next character in the same word, so an ordinary (string-set! s 0
  ...) or (list->string ...) silently corrupted a character the program
  never wrote to, and NTH's own masking hid it at the character that was
  written (C-1).
- EXIT passes the whole 0-255 range through instead of collapsing
  everything above 10.  %HALT's argument is the HALT instruction's 8 bit
  ARG-FIELD, so it has to be a compile time literal; a LOCAL-SYNTAX
  writes out the dispatch.  The out-of-range complaint goes to standard
  error, where diagnostics belong (H-2).
- A switch whose action signals is counted, and --exit and the EOF exit
  out of a non-interactive REPL report a nonzero status, so a make rule
  no longer sees exit 0 next to the stale .oa a failed --compile left in
  place (M-1).
- MAKE-STRING and MAKE-VECTOR take the optional fill R4RS specifies, and
  MAKE-STRING rejects a non-character fill rather than taking apart its
  data field (L-1).
- INTEGER->CHAR rejects a code outside 0-255, which is the range a
  byte-packed string can hold (L-5).
- --pthreads on an emulator without thread support warns and lets the
  switches after it run, rather than warning that it is ignored and then
  terminating (L-3).

The cold linker's new range check turned up a real 32-bit defect:
print-integer.oak's ninth-digit bound was the literal 1000000000, which
is a bignum in a 30-bit-fixnum world.  The same .oa builds both cold
worlds, so the linker had been wrapping it into a negative fixnum and
the nine digit fast path never fired on 32-bit.  Written as a division
instead; it was the only out-of-range literal in the 194 cold objects,
and oak-cold-linker --32bit now links cleanly.

The man page gains an EXIT STATUS section, since the status is now a
contract worth stating.

"make check-cold-linker" in src/world passes, as does the src/misc
suite.

+++ src/emulator/cmdline.h
@@ -24,8 +24,10 @@
 #ifndef _CMDLINE_H_INCLUDED
 #define _CMDLINE_H_INCLUDED
 
+#include <sys/types.h>
+
 extern void parse_cmd_line(int argc, char **argv);
-extern int program_arg_char(int arg_index, int char_index);
+extern int program_arg_char(ssize_t arg_index, ssize_t char_index);
 extern void clear_program_args(void);
 
 #endif
